+++ beacon-chain/sync/service.go
@@ -21,6 +21,7 @@ import (
 	"github.com/prysmaticlabs/prysm/v5/async/abool"
 	"github.com/prysmaticlabs/prysm/v5/async/event"
 	"github.com/prysmaticlabs/prysm/v5/beacon-chain/blockchain"
+	"github.com/prysmaticlabs/prysm/v5/beacon-chain/cache"
 	blockfeed "github.com/prysmaticlabs/prysm/v5/beacon-chain/core/feed/block"
 	"github.com/prysmaticlabs/prysm/v5/beacon-chain/core/feed/operation"
 	statefeed "github.com/prysmaticlabs/prysm/v5/beacon-chain/core/feed/state"
@@ -84,6 +85,7 @@ type config struct {
 	attestationNotifier     operation.Notifier
 	p2p                     p2p.P2P
 	beaconDB                db.NoHeadAccessDatabase
+	attestationCache        *cache.AttestationCache
 	attPool                 attestations.Pool
 	exitPool                voluntaryexits.PoolManager
 	slashingPool            slashings.PoolManager