+++ config/features/flags.go
@@ -174,6 +174,10 @@ var (
 		Name:  "enable-discovery-reboot",
 		Usage: "Experimental: Enables the discovery listener to rebooted in the event of connectivity issues.",
 	}
+	enableExperimentalAttestationPool = &cli.BoolFlag{
+		Name:  "enable-experimental-attestation-pool",
+		Usage: "Enables an experimental attestation pool design.",
+	}
 )
 
 // devModeFlags holds list of flags that are set when development mode is on.
@@ -231,6 +235,7 @@ var BeaconChainFlags = append(deprecatedBeaconFlags, append(deprecatedFlags, []c
 	DisableQUIC,
 	DisableCommitteeAwarePacking,
 	EnableDiscoveryReboot,
+	enableExperimentalAttestationPool,
 }...)...)
 
 // E2EBeaconChainFlags contains a list of the beacon chain feature flags to be tested in E2E.